requirements.txt
flask==3.1.0
flask_socketio==5.5.1
docker==7.1.0
python-magic==0.4.27
pyrebase4
bleach
python-dotenv
